+++ tests/Feature/Api/V1/ItemsTest.php
@@ -4,18 +4,20 @@ namespace Tests\Feature\Api\V1;
 
 use App\Elasticsearch\Repositories\ItemRepository;
 use App\Item;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\RecreateSearchIndex;
 use Tests\TestCase;
 
 class ItemsTest extends TestCase
 {
+    use RefreshDatabase;
     use RecreateSearchIndex;
 
     public function setUp(): void
     {
         parent::setUp();
 
-        Item::factory()->create(['topic' => 'spring']);
+        Item::factory()->create(['topic' => 'spring', 'has_image' => true]);
         Item::factory()->create(['topic' => 'summer']);
         Item::factory()->create(['topic' => 'autumn']);
 
@@ -51,6 +53,17 @@ class ItemsTest extends TestCase
 
         $this->getJson($url)->assertJson([
             'total' => 2,
+        ]);
+    }
+
+    public function test_filtering_by_boolean_values()
+    {
+        $url = route('api.v1.items.index', [
+            'filter[has_image]' => 'true',
+        ]);
+
+        $this->getJson($url)->assertJson([
+            'total' => 1,
         ]);
     }
 }