+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [Unreleased]
 ### Added
 - multi-term filtering to /api/v1/items
+- support for boolean filter values in /api/v1/items
 
 ## [2.72.1] - 2023-02-10
 ### Added